requirements_databroker.txt
PyMySQL==1.1.0
python-dotenv==0.19.0
watchdog==2.1.6